Change playRound and playGame to handle draws
Changed playRound and playGame to handle draws. Moved the declaration
and initialization of humanChoice and computerChoice to playRound and
removed them as parameters. Changed the if statement checking for draws
to a while statement and re-assigned humanChoice and computerChoice
while they are equal. Also removed pseudocode from the top of the
program