Move script tag down, add click event listener
Moved the script tag to the bottom of index.html, added the button
click event listener to receive the user choice and play a round.
Removed the now redundant getHumanChoice and playGame functions. Also
fixed logic to handle a draw in compareChoice by removing the call to
the getHumanChoice function and adding a break, otherwise an infinite
loop was encountered.