Change menu ID to a class and add more styles
Changed the menu ID to a class and refactored the selector of it in
javascript.js. Also moved the restartGame and whoWon functions into
the playGame function to avoid global variables. Added several div's and
classes to index.html to work with new styles in style.css. Added new
font rules.